b/f: UJObjectUnpack now respects output sequence
The function tries to match decoded JSON objects to the given keys
received as arguments and store the corresponding objects back into the
out parameters.  The bug stems from the fact that walking the list of va
arguments is synchronised with the iteration over the members of the
decoded object, not the keys. Since the order of the members into a JSON
object can vary, the allocation to the output references can be wrong.

Also, the usage of va_arg in the function receives the wrong type
paramter: UJObject instead of UJObject *, though this has little
consequence, since both are pointers.